fix: Correct the usage of CAM argument input
When the argument -i (--input) is passed as CAM or 0, the code changes input as 0 (integer)
in order to indicate to OpenCV that it needs to open the WebCam.
As in the next lines the code searched for the file extension with .endswith(), it didn't
worked when the argument was CAM or 0. Fixed by changing the order and indentation.